code/segmentation/requirements.txt
-e git+https://github.com/facebookresearch/detectron2.git#egg=detectron2
matplotlib==3.7.5
numpy==1.24.4
opencv_python==4.8.1.78
pandas==2.0.3
pycocotools==2.0.8
PyYAML==6.0.1
PyYAML==6.0.2
scikit_learn==1.3.2
torch==2.3.0
torchvision==0.18.0
tqdm==4.66.4
